Added recursion and threading problems
Made call to SearchForIDs recursive so that it will keep calling until all consignment Ids are processed.
problem where SearchForIDs can be called by two threads at the same time causing ConsignmentIdIndex to do weird things and CEF is not going to be too happy about it either.
Added locknig to try and get around that. this is only first attempt.